upgrade tests to use TypeScript
