+++ test/karma.config.cjs
@@ -1,5 +1,3 @@
-const resolve = require('@rollup/plugin-node-resolve').default
-
 process.env.CHROME_BIN = require('chromium').path
 
 module.exports = function (config) {
@@ -9,18 +7,13 @@ module.exports = function (config) {
     files: [
       {pattern: 'lib/*.js', type: 'module', included: false},
       {pattern: 'node_modules/**', type: 'module', included: false},
-      {pattern: 'test/*', type: 'module', included: true, watched: false}
+      {pattern: 'test/*.ts', type: 'module', included: true, watched: false}
     ],
     preprocessors: {
-      'test/*.ts': ['rollup']
+      'test/*.ts': ['esbuild']
     },
-    rollupPreprocessor: {
-      plugins: [resolve()],
-      output: {
-        format: 'iife',
-        name: 'test',
-        sourcemap: 'inline'
-      }
+    esbuild: {
+      target: 'es2019'
     },
     reporters: ['mocha'],
     port: 9876,
